backend/requirements.txt
asgiref==3.5.2
botometer
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
distlib==0.3.4
Django==4.0.6
django-cuser==2017.3.16
filelock==3.7.0
fonttools==4.38.0
google-auth==2.13.0
google-auth-oauthlib==0.6.0
gspread==5.6.2
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
pbr==5.9.0
Pillow==9.3.0
platformdirs==2.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scripts==2.0
six==1.16.0
sqlparse==0.4.2
stevedore==3.5.0
stua==0.2
tweepy==4.12.1
urllib3==1.26.12
virtualenv==20.14.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
